We add style to our JSX elements, in this case we add a border to our header
We use inline style or className. We inject the style object using {}

Every CSS property becomes a key and every CSS property value becomes a value for the the object. For instance, in our case, color is a key and 'black' is a value, fontSize is a key and '18px' is a value. All two word CSS properties will change to camelCase when we use them as key in the CSS object in React